Autthorization Added

+++ backend/routes/posts.js
@@ -37,6 +37,7 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
         title: req.body.title,
         content: req.body.content,
         imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId,
     });
     post.save().then(result => {
         console.log(post);
@@ -64,12 +65,17 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
     console.log("!!!!!!!!" + post);
-    PostModel.updateOne({_id: req.params.id}, post).then(result => {
-        console.log(result);
-        res.status(200).json({message: 'Update post successful'});
+    PostModel.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if(result.nModified > 0) {
+            res.status(201).json({ message: "Update successful"});
+
+        } else {
+            res.status(401).json({ message: "User not authorized"});
+        }
     });
 });
 
@@ -114,11 +120,13 @@ router.get("/:id", (req, res, next) => {
 
 // DELETE POST BY ID
 router.delete("/:id", checkAuth, (req, res, next) => {
-    PostModel.deleteOne({_id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({
-            message: 'Post deleted.',
-        });
+    PostModel.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if(result.n > 0) {
+            res.status(201).json({ message: "Update successful"});
+
+        } else {
+            res.status(401).json({ message: "User not authorized"});
+        }
     });
